refactor(CoinService): dependency injection for coin logic
