+++ src/main/java/vendingmachine/domain/VendingMachine.java
@@ -4,6 +4,7 @@ import static vendingmachine.constants.SystemConstants.*;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import vendingmachine.domain.enums.Coin;
+import vendingmachine.service.CoinService;
 import vendingmachine.view.InputView;
 import vendingmachine.view.OutputView;
 
@@ -11,12 +12,12 @@ import java.util.HashMap;
 
 public class VendingMachine {
 
-    private final HashMap<Coin, Integer> coins = new HashMap<>();
+    private final CoinService coinService = new CoinService();
     private Menu menu;
     private int customerMoneyLeft = NO_CUSTOMER_MONEY_LEFT;
 
     public VendingMachine() {
-        this.initializeCoins(InputView.getInitialMoneyInput());
+        coinService.initializeCoins(InputView.getInitialMoneyInput());
         this.setMenu(InputView.getMenuInput());
         this.setCustomerMoneyLeft(InputView.getCustomerMoneyInput());
     }
@@ -30,26 +31,6 @@ public class VendingMachine {
         }
 
         OutputView.printCoinChanges(this.calculateCoinChanges());
-    }
-
-    private void initializeCoins(int totalMoney) {
-        for (Coin coin : Coin.values()) {
-            this.coins.put(coin, ZERO_COINS);
-        }
-        generateCoins(totalMoney);
-
-        OutputView.printCoinsInfo(this.coins);
-    }
-
-    private void generateCoins(int totalMoney) {
-        while (totalMoney > NO_TOTAL_MONEY_LEFT) {
-            int randomCoinAmount = Randoms.pickNumberInList(Coin.getCoinsAmountDesc());
-            if (totalMoney >= randomCoinAmount) {
-                Coin newCoin = Coin.getCoinByAmount(randomCoinAmount);
-                coins.put(newCoin, coins.get(newCoin) + 1);
-                totalMoney -= randomCoinAmount;
-            }
-        }
     }
 
     private void setMenu(Menu menu) {
@@ -73,11 +54,11 @@ public class VendingMachine {
     private HashMap<Coin, Integer> calculateCoinChanges() {
         HashMap<Coin, Integer> coinChanges = new HashMap<>();
         for (Coin coin : Coin.getCoinsDesc()) {
-            if (this.coins.get(coin) == ZERO_COINS) continue;
+            if (this.coinService.getCoins().get(coin) == ZERO_COINS) continue;
             if (this.customerMoneyLeft < coin.getAmount()) continue;
             int coinNumber = this.calculateMaximumCoinNumber(coin);
             coinChanges.put(coin, coinNumber);
-            this.coins.put(coin, this.coins.get(coin)-coinNumber);
+            this.coinService.getCoins().put(coin, this.coinService.getCoins().get(coin)-coinNumber);
             this.customerMoneyLeft -= coin.getAmount() * coinNumber;
         }
         return coinChanges;
@@ -85,8 +66,8 @@ public class VendingMachine {
 
     private int calculateMaximumCoinNumber(Coin coin) {
         int spentCoinNumber = this.customerMoneyLeft/coin.getAmount();
-        if (spentCoinNumber > this.coins.get(coin)) {
-            spentCoinNumber = this.coins.get(coin);
+        if (spentCoinNumber > this.coinService.getCoins().get(coin)) {
+            spentCoinNumber = this.coinService.getCoins().get(coin);
         }
         return spentCoinNumber;
     }
